my-project-backend/backend/requirements.txt
asgiref==3.11.0
distlib==0.4.0
dj-database-url==3.1.0
Django==5.2.9
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
filelock==3.20.0
gunicorn==25.0.1
packaging==26.0
pillow==12.0.0
platformdirs==4.5.0
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg2-binary==2.9.11
sqlparse==0.5.4
tzdata==2025.2
virtualenv==20.35.4
whitenoise==6.11.0
